+++ NEWS.md
@@ -1,5 +1,8 @@
-# bbplot 0.0.1
+# plotbb 0.0.1
 
++ `as.bbplot` to convert function as `bbplot` object
++ `bb_grid` to add grid line (background grid line by default)
++ `bb_theme_expand`, `bb_theme_grey` and `bb_theme_deepblue`
 + `+` operator supported, e.g. `bbplot() + bb_point()`
 + `bb_theme` for setting graphic parameters
 + `bb_point` and `bb_lm` geometric layers